replace karts batching with srb2s batching
currently palette renderer does not work
may need to try it with the reworked
