+++ src/hardware/hw_drv.h
@@ -36,7 +36,8 @@ EXPORT void HWRAPI(SetPalette) (RGBA_t *ppal, RGBA_t *pgamma);
 #endif
 EXPORT void HWRAPI(FinishUpdate) (INT32 waitvbl);
 EXPORT void HWRAPI(Draw2DLine) (F2DCoord *v1, F2DCoord *v2, RGBA_t Color);
-EXPORT void HWRAPI(DrawPolygon) (FSurfaceInfo *pSurf, FOutVector *pOutVerts, FUINT iNumPts, FBITFIELD PolyFlags, boolean horizonSpecial);
+EXPORT void HWRAPI(DrawPolygon) (FSurfaceInfo *pSurf, FOutVector *pOutVerts, FUINT iNumPts, FBITFIELD PolyFlags);
+EXPORT void HWRAPI(DrawIndexedTriangles) (FSurfaceInfo *pSurf, FOutVector *pOutVerts, FUINT iNumPts, FBITFIELD PolyFlags, unsigned int *IndexArray);
 EXPORT void HWRAPI(SetBlend) (FBITFIELD PolyFlags);
 EXPORT void HWRAPI(ClearBuffer) (FBOOLEAN ColorMask, FBOOLEAN DepthMask, FBOOLEAN StencilMask, FRGBAFloat *ClearColor);
 EXPORT void HWRAPI(SetTexture) (FTextureInfo *TexInfo);
@@ -81,9 +82,6 @@ EXPORT void HWRAPI(SetShaderInfo) (hwdshaderinfo_t info, INT32 value);
 EXPORT void HWRAPI(LoadCustomShader) (int number, char *shader, size_t size, boolean fragment);
 EXPORT boolean HWRAPI(InitCustomShaders) (void);
 
-EXPORT void HWRAPI(StartBatching) (void);
-EXPORT void HWRAPI(RenderBatches) (precise_t *sSortTime, precise_t *sDrawTime, int *sNumPolys, int *sNumVerts, int *sNumCalls, int *sNumShaders, int *sNumTextures, int *sNumPolyFlags, int *sNumColors);
-
 EXPORT void HWRAPI(InitPalette) (int flashnum, boolean skiplut);
 EXPORT UINT32 HWRAPI(AddLightTable) (UINT8 *lighttable);
 EXPORT void HWRAPI(ClearLightTableCache) (void);
@@ -102,6 +100,7 @@ struct hwdriver_s
 	FinishUpdate        	pfnFinishUpdate;
 	Draw2DLine          	pfnDraw2DLine;
 	DrawPolygon         	pfnDrawPolygon;
+	DrawIndexedTriangles    pfnDrawIndexedTriangles;
 	SetBlend            	pfnSetBlend;
 	ClearBuffer         	pfnClearBuffer;
 	SetTexture          	pfnSetTexture;
@@ -136,9 +135,6 @@ struct hwdriver_s
 	LoadCustomShader 		pfnLoadCustomShader;
 	InitCustomShaders 		pfnInitCustomShaders;
 
-	StartBatching 			pfnStartBatching;
-	RenderBatches 			pfnRenderBatches;
-	
 	InitPalette 			pfnInitPalette;
 	AddLightTable 			pfnAddLightTable;
 	ClearLightTableCache 	pfnClearLightTableCache;
