+++ src/hardware/r_opengl/r_opengl.c
@@ -34,6 +34,7 @@
 #include "../../i_system.h" // for I_GetPreciseTime (batching time measurements)
 
 #include "../hw_main.h"
+#include "../hw_batching.h"
 
 // Eeeeh not sure is this right way, but it works < sry :c 
 extern consvar_t cv_grusecustomshaders;
@@ -637,8 +638,6 @@ static boolean gl_allowshaders = false;
 static boolean gl_shadersenabled = false;
 static GLuint gl_currentshaderprogram = 0;
 static boolean gl_shaderprogramchanged = true;
-
-static boolean gl_batching = false;// are we currently collecting batches?
 
 static GLint gl_palette[768];
 boolean gl_palette_initialized = false;
@@ -1464,7 +1463,7 @@ EXPORT void HWRAPI(InitPalette) (int flashnum, boolean skiplut)
 static void SetNoTexture(void)
 {
 	// Disable texture.
-	if (tex_downloaded != NOTEXTURE_NUM && !gl_batching)
+	if (tex_downloaded != NOTEXTURE_NUM)
 	{
 		pglBindTexture(GL_TEXTURE_2D, NOTEXTURE_NUM);
 		tex_downloaded = NOTEXTURE_NUM;
@@ -1685,7 +1684,7 @@ EXPORT UINT32 HWRAPI(AddLightTable) (UINT8 *lighttable)
 	pglTexImage2D(GL_TEXTURE_2D, 0, internalFormat, 256, 32, 0, GL_RED, GL_UNSIGNED_BYTE, lighttable);
 
 	// restore previously bound texture
-	if (!gl_batching)
+	if (!cv_grbatching.value)
 		pglBindTexture(GL_TEXTURE_2D, tex_downloaded);
 
 	return ltcachetail->id;
@@ -2362,628 +2361,77 @@ static void Shader_SetUniforms(FSurfaceInfo *Surface, GLRGBAFloat *poly, GLRGBAF
 #endif
 }
 
-
-// unfinished draw call batching
-
-// Note: could use realloc in the array reallocation code
-
-typedef struct
+// code that is common between DrawPolygon and DrawIndexedTriangles
+// the corona thing is there too, i have no idea if that stuff works with DrawIndexedTriangles and batching
+static void PreparePolygon(FSurfaceInfo *pSurf, FOutVector *pOutVerts, FBITFIELD PolyFlags)
 {
-	FSurfaceInfo surf;// surf also has its own polyflags for some reason, but it seems unused
-	unsigned int vertsIndex;// location of verts in unsortedVertexArray
-	FUINT numVerts;
-	FBITFIELD polyFlags;
-	GLuint texNum;
-	GLuint shader;
-	// this tells batching that the plane belongs to a horizon line and must be drawn in correct order with the skywalls
-	boolean horizonSpecial;
-} PolygonArrayEntry;
+	static GLRGBAFloat poly = {0,0,0,0};
+	static GLRGBAFloat tint = {0,0,0,0};
+	static GLRGBAFloat fade = {0,0,0,0};
 
-FOutVector* finalVertexArray = NULL;// contains subset of sorted vertices and texture coordinates to be sent to gpu
-UINT32* finalVertexIndexArray = NULL;// contains indexes for glDrawElements, taking into account fan->triangles conversion
-//     NOTE have this alloced as 3x finalVertexArray size
-int finalVertexArrayAllocSize = 65536;
-//GLubyte* colorArray = NULL;// contains color data to be sent to gpu, if needed
-//int colorArrayAllocSize = 65536;
-// not gonna use this for now, just sort by color and change state when it changes
-// later maybe when using vertex attributes if it's needed
+	SetBlend(PolyFlags);    //TODO: inline (#pragma..)
 
-PolygonArrayEntry* polygonArray = NULL;// contains the polygon data from DrawPolygon, waiting to be processed
-int polygonArraySize = 0;
-UINT32* polygonIndexArray = NULL;// contains sorting pointers for polygonArray
-int polygonArrayAllocSize = 65536;
-
-FOutVector* unsortedVertexArray = NULL;// contains unsorted vertices and texture coordinates from DrawPolygon
-int unsortedVertexArraySize = 0;
-int unsortedVertexArrayAllocSize = 65536;
-
-EXPORT void HWRAPI(StartBatching) (void)
-{
-	//CONS_Printf("StartBatching begin\n");
-	// init arrays if that has not been done yet
-	if (!finalVertexArray)
+	// PolyColor
+	if (pSurf)
 	{
-		finalVertexArray = malloc(finalVertexArrayAllocSize * sizeof(FOutVector));
-		finalVertexIndexArray = malloc(finalVertexArrayAllocSize * 3 * sizeof(UINT32));
-		polygonArray = malloc(polygonArrayAllocSize * sizeof(PolygonArrayEntry));
-		polygonIndexArray = malloc(polygonArrayAllocSize * sizeof(UINT32));
-		unsortedVertexArray = malloc(unsortedVertexArrayAllocSize * sizeof(FOutVector));
-	}
-	// drawing functions will now collect the drawing data instead of passing it to opengl
-	gl_batching = true;
-	//CONS_Printf("StartBatching end\n");
-}
+		// If Modulated, mix the surface colour to the texture
+		if (CurrentPolyFlags & PF_Modulated)
+		{
+			// Poly color
+			poly.red    = byte2float[pSurf->PolyColor.s.red];
+			poly.green  = byte2float[pSurf->PolyColor.s.green];
+			poly.blue   = byte2float[pSurf->PolyColor.s.blue];
+			poly.alpha  = byte2float[pSurf->PolyColor.s.alpha];
 
-static int comparePolygons(const void *p1, const void *p2)
-{
-	unsigned int index1 = *(const unsigned int*)p1;
-	unsigned int index2 = *(const unsigned int*)p2;
-	PolygonArrayEntry* poly1 = &polygonArray[index1];
-	PolygonArrayEntry* poly2 = &polygonArray[index2];
-	int diff;
-	INT64 diff64;
+			pglColor4ubv((GLubyte*)&pSurf->PolyColor.s);
+		}
 
-	int shader1 = poly1->shader;
-	int shader2 = poly2->shader;
-	// make skywalls and horizon lines first in order
-	if (poly1->polyFlags & PF_NoTexture || poly1->horizonSpecial)
-		shader1 = -1;
-	if (poly2->polyFlags & PF_NoTexture || poly2->horizonSpecial)
-		shader2 = -1;
-	diff = shader1 - shader2;
-	if (diff != 0) return diff;
+		// Tint color
+		tint.red   = byte2float[pSurf->TintColor.s.red];
+		tint.green = byte2float[pSurf->TintColor.s.green];
+		tint.blue  = byte2float[pSurf->TintColor.s.blue];
+		tint.alpha = byte2float[pSurf->TintColor.s.alpha];
 
-	// skywalls and horizon lines must retain their order for horizon lines to work
-	if (shader1 == -1 && shader2 == -1)
-		return index1 - index2;
-
-	diff = poly1->texNum - poly2->texNum;
-	if (diff != 0) return diff;
-	
-	if (gl_palshader)
-	{
-		diff = poly1->surf.LightTableId - poly2->surf.LightTableId;
-		if (diff != 0) return diff;
+		// Fade color
+		fade.red   = byte2float[pSurf->FadeColor.s.red];
+		fade.green = byte2float[pSurf->FadeColor.s.green];
+		fade.blue  = byte2float[pSurf->FadeColor.s.blue];
+		fade.alpha = byte2float[pSurf->FadeColor.s.alpha];
 	}
 
-	diff = poly1->polyFlags - poly2->polyFlags;
-	if (diff != 0) return diff;
-
-	diff64 = poly1->surf.PolyColor.rgba - poly2->surf.PolyColor.rgba;
-	if (diff64 < 0) return -1; else if (diff64 > 0) return 1;
-	diff64 = poly1->surf.TintColor.rgba - poly2->surf.TintColor.rgba;
-	if (diff64 < 0) return -1; else if (diff64 > 0) return 1;
-	diff64 = poly1->surf.FadeColor.rgba - poly2->surf.FadeColor.rgba;
-	if (diff64 < 0) return -1; else if (diff64 > 0) return 1;
-
-	diff = poly1->surf.LightInfo.light_level - poly2->surf.LightInfo.light_level;
-	if (diff != 0) return diff;
-	diff = poly1->surf.LightInfo.fade_start - poly2->surf.LightInfo.fade_start;
-	if (diff != 0) return diff;
-	diff = poly1->surf.LightInfo.fade_end - poly2->surf.LightInfo.fade_end;
-	return diff;
-}
-
-static int comparePolygonsNoShaders(const void *p1, const void *p2)
-{
-	unsigned int index1 = *(const unsigned int*)p1;
-	unsigned int index2 = *(const unsigned int*)p2;
-	PolygonArrayEntry* poly1 = &polygonArray[index1];
-	PolygonArrayEntry* poly2 = &polygonArray[index2];
-	int diff;
-	INT64 diff64;
-
-	GLuint texNum1 = poly1->texNum;
-	GLuint texNum2 = poly2->texNum;
-	if (poly1->polyFlags & PF_NoTexture || poly1->horizonSpecial)
-		texNum1 = 0;
-	if (poly2->polyFlags & PF_NoTexture || poly2->horizonSpecial)
-		texNum2 = 0;
-	diff = texNum1 - texNum2;
-	if (diff != 0) return diff;
-
-	// skywalls and horizon lines must retain their order for horizon lines to work
-	if (texNum1 == 0 && texNum2 == 0)
-		return index1 - index2;
-
-	diff = poly1->polyFlags - poly2->polyFlags;
-	if (diff != 0) return diff;
-
-	diff64 = poly1->surf.PolyColor.rgba - poly2->surf.PolyColor.rgba;
-	if (diff64 < 0) return -1; else if (diff64 > 0) return 1;
-
-	return 0;
-}
-
-// the parameters for this functions (numPolys etc.) are used to return rendering stats
-EXPORT void HWRAPI(RenderBatches) (precise_t *sSortTime, precise_t *sDrawTime, int *sNumPolys, int *sNumVerts, int *sNumCalls, int *sNumShaders, int *sNumTextures, int *sNumPolyFlags, int *sNumColors)
-{
-	int finalVertexWritePos = 0;// position in finalVertexArray
-	int finalIndexWritePos = 0;// position in finalVertexIndexArray
-
-	int polygonReadPos = 0;// position in polygonIndexArray
-
-	GLuint currentShader;
-	GLuint currentTexture;
-	FBITFIELD currentPolyFlags;
-	FSurfaceInfo currentSurfaceInfo;
-
-	GLRGBAFloat firstPoly = {0,0,0,0}; // may be misleading but this means first PolyColor
-	GLRGBAFloat firstTint = {0,0,0,0};
-	GLRGBAFloat firstFade = {0,0,0,0};
-
-	boolean needRebind = false;
-
-	int i;
-
-	//CONS_Printf("RenderBatches\n");
-	gl_batching = false;// no longer collecting batches
-	if (!polygonArraySize)
-	{
-		*sSortTime = *sDrawTime = *sNumPolys = *sNumCalls = *sNumShaders = *sNumTextures = *sNumPolyFlags = *sNumColors = 0;
-		return;// nothing to draw
-	}
-	// init stats vars
-	*sNumPolys = polygonArraySize;
-	*sNumCalls = *sNumVerts = 0;
-	*sNumShaders = *sNumTextures = *sNumPolyFlags = *sNumColors = 1;
-	// init polygonIndexArray
-	for (i = 0; i < polygonArraySize; i++)
-	{
-		polygonIndexArray[i] = i;
-	}
-
-	// sort polygons
-	//CONS_Printf("qsort polys\n");
-	*sSortTime = I_GetPreciseTime();
-	if (gl_allowshaders)
-		qsort(polygonIndexArray, polygonArraySize, sizeof(unsigned int), comparePolygons);
-	else
-		qsort(polygonIndexArray, polygonArraySize, sizeof(unsigned int), comparePolygonsNoShaders);
-	*sSortTime = I_GetPreciseTime() - *sSortTime;
-	//CONS_Printf("sort done\n");
-	// sort order
-	// 1. shader
-	// 2. texture
-	// 3. polyflags
-	// 4. colors + light level
-	// not sure about order of last 2, or if it even matters
-
-	*sDrawTime = I_GetPreciseTime();
-
-	currentShader = polygonArray[polygonIndexArray[0]].shader;
-	currentTexture = polygonArray[polygonIndexArray[0]].texNum;
-	currentPolyFlags = polygonArray[polygonIndexArray[0]].polyFlags;
-	currentSurfaceInfo = polygonArray[polygonIndexArray[0]].surf;
-	// For now, will sort and track the colors. Vertex attributes could be used instead of uniforms
-	// and a color array could replace the color calls.
-
-	// set state for first batch
-	//CONS_Printf("set first state\n");
-	gl_currentshaderprogram = currentShader;
-	gl_shaderprogramchanged = true;
-	if (currentPolyFlags & PF_Modulated)
-	{
-		// Poly color
-		firstPoly.red    = byte2float[currentSurfaceInfo.PolyColor.s.red];
-		firstPoly.green  = byte2float[currentSurfaceInfo.PolyColor.s.green];
-		firstPoly.blue   = byte2float[currentSurfaceInfo.PolyColor.s.blue];
-		firstPoly.alpha  = byte2float[currentSurfaceInfo.PolyColor.s.alpha];
-		pglColor4ubv((GLubyte*)&currentSurfaceInfo.PolyColor.s);
-	}
-	// Tint color
-	firstTint.red   = byte2float[currentSurfaceInfo.TintColor.s.red];
-	firstTint.green = byte2float[currentSurfaceInfo.TintColor.s.green];
-	firstTint.blue  = byte2float[currentSurfaceInfo.TintColor.s.blue];
-	firstTint.alpha = byte2float[currentSurfaceInfo.TintColor.s.alpha];
-	// Fade color
-	firstFade.red   = byte2float[currentSurfaceInfo.FadeColor.s.red];
-	firstFade.green = byte2float[currentSurfaceInfo.FadeColor.s.green];
-	firstFade.blue  = byte2float[currentSurfaceInfo.FadeColor.s.blue];
-	firstFade.alpha = byte2float[currentSurfaceInfo.FadeColor.s.alpha];
-
-	if (gl_allowshaders)
-	{
-		Shader_Load(&currentSurfaceInfo, &firstPoly, &firstTint, &firstFade);
-		if (gl_palshader)
-		{
-			pglActiveTexture(GL_TEXTURE2);// this stuff could be done better but gonna do it quick like this for now
-			pglBindTexture(GL_TEXTURE_2D, currentSurfaceInfo.LightTableId);
-			pglActiveTexture(GL_TEXTURE0);
-		}
-	}
-
-	if (currentPolyFlags & PF_NoTexture)
-		currentTexture = 0;
-	pglBindTexture(GL_TEXTURE_2D, currentTexture);
-	tex_downloaded = currentTexture;
-
-	SetBlend(currentPolyFlags);
-
-	//CONS_Printf("first pointers to ogl\n");
-	pglVertexPointer(3, GL_FLOAT, sizeof(FOutVector), &finalVertexArray[0].x);
-	pglTexCoordPointer(2, GL_FLOAT, sizeof(FOutVector), &finalVertexArray[0].s);
-
-	while (1)// note: remember handling notexture polyflag as having texture number 0 (also in comparePolygons)
-	{
-		int firstIndex;
-		int lastIndex;
-
-		boolean stopFlag = false;
-		boolean changeState = false;
-		boolean changeShader = false;
-		GLuint nextShader = 0U;
-		boolean changeTexture = false;
-		GLuint nextTexture = 0U;
-		boolean changePolyFlags = false;
-		FBITFIELD nextPolyFlags = 0U;
-		boolean changeSurfaceInfo = false;
-		FSurfaceInfo nextSurfaceInfo;
-
-		//CONS_Printf("loop iter start\n");
-		// new try:
-		// write vertices
-		// check for changes or end, otherwise go back to writing
-			// changes will affect the next vars and the change bools
-			// end could set flag for stopping
-		// execute draw call
-		// could check ending flag here
-		// change states according to next vars and change bools, updating the current vars and reseting the bools
-		// reset write pos
-		// repeat loop
-
-		int index = polygonIndexArray[polygonReadPos++];
-		int numVerts = polygonArray[index].numVerts;
-		// before writing, check if there is enough room
-		// using 'while' instead of 'if' here makes sure that there will *always* be enough room.
-		// probably never will this loop run more than once though
-		while (finalVertexWritePos + numVerts > finalVertexArrayAllocSize)
-		{
-			FOutVector* new_array;
-			unsigned int* new_index_array;
-			//CONS_Printf("final vert realloc\n");
-			finalVertexArrayAllocSize *= 2;
-			new_array = malloc(finalVertexArrayAllocSize * sizeof(FOutVector));
-			memcpy(new_array, finalVertexArray, finalVertexWritePos * sizeof(FOutVector));
-			free(finalVertexArray);
-			finalVertexArray = new_array;
-			// also increase size of index array, 3x of vertex array since
-			// going from fans to triangles increases vertex count to 3x
-			new_index_array = malloc(finalVertexArrayAllocSize * 3 * sizeof(UINT32));
-			memcpy(new_index_array, finalVertexIndexArray, finalIndexWritePos * sizeof(UINT32));
-			free(finalVertexIndexArray);
-			finalVertexIndexArray = new_index_array;
-			// if vertex buffers are reallocated then opengl needs to know too
-			needRebind = true;
-		}
-		//CONS_Printf("write verts to final\n");
-		// write the vertices of the polygon
-		memcpy(&finalVertexArray[finalVertexWritePos], &unsortedVertexArray[polygonArray[index].vertsIndex],
-			numVerts * sizeof(FOutVector));
-		// write the indexes, pointing to the fan vertexes but in triangles format
-		firstIndex = finalVertexWritePos;
-		lastIndex = finalVertexWritePos + numVerts;
-		finalVertexWritePos += 2;
-		//CONS_Printf("write final vert indices\n");
-		while (finalVertexWritePos < lastIndex)
-		{
-			finalVertexIndexArray[finalIndexWritePos++] = firstIndex;
-			finalVertexIndexArray[finalIndexWritePos++] = finalVertexWritePos - 1;
-			finalVertexIndexArray[finalIndexWritePos++] = finalVertexWritePos++;
-		}
-
-		if (polygonReadPos >= polygonArraySize)
-		{
-			stopFlag = true;
-		}
-		else
-		{
-			//CONS_Printf("state change check\n");
-			// check if a state change is required, set the change bools and next vars
-			int nextIndex = polygonIndexArray[polygonReadPos];
-			nextShader = polygonArray[nextIndex].shader;
-			nextTexture = polygonArray[nextIndex].texNum;
-			nextPolyFlags = polygonArray[nextIndex].polyFlags;
-			nextSurfaceInfo = polygonArray[nextIndex].surf;
-			if (nextPolyFlags & PF_NoTexture)
-				nextTexture = 0;
-			if (currentShader != nextShader)
-			{
-				changeState = true;
-				changeShader = true;
-			}
-			if (currentTexture != nextTexture)
-			{
-				changeState = true;
-				changeTexture = true;
-			}
-			if (currentPolyFlags != nextPolyFlags)
-			{
-				changeState = true;
-				changePolyFlags = true;
-			}
-			if (gl_allowshaders && gl_palshader)
-			{
-				if (currentSurfaceInfo.PolyColor.rgba != nextSurfaceInfo.PolyColor.rgba ||
-					currentSurfaceInfo.TintColor.rgba != nextSurfaceInfo.TintColor.rgba ||
-					currentSurfaceInfo.FadeColor.rgba != nextSurfaceInfo.FadeColor.rgba ||
-					currentSurfaceInfo.LightInfo.light_level != nextSurfaceInfo.LightInfo.light_level ||
-					currentSurfaceInfo.LightInfo.fade_start != nextSurfaceInfo.LightInfo.fade_start ||
-					currentSurfaceInfo.LightInfo.fade_end != nextSurfaceInfo.LightInfo.fade_end ||
-					currentSurfaceInfo.LightTableId != nextSurfaceInfo.LightTableId)
-				{
-					changeState = true;
-					changeSurfaceInfo = true;
-				}
-			}
-			else if ((gl_allowshaders) && (!gl_palshader))
-			{
-				if (currentSurfaceInfo.PolyColor.rgba != nextSurfaceInfo.PolyColor.rgba ||
-					currentSurfaceInfo.TintColor.rgba != nextSurfaceInfo.TintColor.rgba ||
-					currentSurfaceInfo.FadeColor.rgba != nextSurfaceInfo.FadeColor.rgba ||
-					currentSurfaceInfo.LightInfo.light_level != nextSurfaceInfo.LightInfo.light_level ||
-					currentSurfaceInfo.LightInfo.fade_start != nextSurfaceInfo.LightInfo.fade_start ||
-					currentSurfaceInfo.LightInfo.fade_end != nextSurfaceInfo.LightInfo.fade_end)
-				{
-					changeState = true;
-					changeSurfaceInfo = true;
-				}
-			}
-			else
-			{
-				if (currentSurfaceInfo.PolyColor.rgba != nextSurfaceInfo.PolyColor.rgba)
-				{
-					changeState = true;
-					changeSurfaceInfo = true;
-				}
-			}
-		}
-
-		if (changeState || stopFlag)
-		{
-			if (needRebind)
-			{
-				//CONS_Printf("rebind\n");
-				pglVertexPointer(3, GL_FLOAT, sizeof(FOutVector), &finalVertexArray[0].x);
-				pglTexCoordPointer(2, GL_FLOAT, sizeof(FOutVector), &finalVertexArray[0].s);
-				needRebind = false;
-			}
-			//CONS_Printf("exec draw call\n");
-			// execute draw call
-			pglDrawElements(GL_TRIANGLES, finalIndexWritePos, GL_UNSIGNED_INT, finalVertexIndexArray);
-			//CONS_Printf("draw call done\n");
-			// update stats
-			(*sNumCalls)++;
-			*sNumVerts += finalIndexWritePos;
-			// reset write positions
-			finalVertexWritePos = 0;
-			finalIndexWritePos = 0;
-		}
-		else continue;
-
-		// if we're here then either its time to stop or time to change state
-		if (stopFlag) break;
-
-		//CONS_Printf("state change\n");
-
-		// change state according to change bools and next vars, update current vars and reset bools
-		if (changeShader)
-		{
-			GLRGBAFloat poly = {0,0,0,0};
-			GLRGBAFloat tint = {0,0,0,0};
-			GLRGBAFloat fade = {0,0,0,0};
-			gl_currentshaderprogram = nextShader;
-			gl_shaderprogramchanged = true;
-			if (nextPolyFlags & PF_Modulated)
-			{
-				// Poly color
-				poly.red    = byte2float[nextSurfaceInfo.PolyColor.s.red];
-				poly.green  = byte2float[nextSurfaceInfo.PolyColor.s.green];
-				poly.blue   = byte2float[nextSurfaceInfo.PolyColor.s.blue];
-				poly.alpha  = byte2float[nextSurfaceInfo.PolyColor.s.alpha];
-			}
-			// Tint color
-			tint.red   = byte2float[nextSurfaceInfo.TintColor.s.red];
-			tint.green = byte2float[nextSurfaceInfo.TintColor.s.green];
-			tint.blue  = byte2float[nextSurfaceInfo.TintColor.s.blue];
-			tint.alpha = byte2float[nextSurfaceInfo.TintColor.s.alpha];
-			// Fade color
-			fade.red   = byte2float[nextSurfaceInfo.FadeColor.s.red];
-			fade.green = byte2float[nextSurfaceInfo.FadeColor.s.green];
-			fade.blue  = byte2float[nextSurfaceInfo.FadeColor.s.blue];
-			fade.alpha = byte2float[nextSurfaceInfo.FadeColor.s.alpha];
-
-			Shader_Load(&nextSurfaceInfo, &poly, &tint, &fade);
-			currentShader = nextShader;
-			changeShader = false;
-
-			(*sNumShaders)++;
-		}
-		if (changeTexture)
-		{
-			// texture should be already ready for use from calls to SetTexture during batch collection
-			pglBindTexture(GL_TEXTURE_2D, nextTexture);
-			tex_downloaded = nextTexture;
-			currentTexture = nextTexture;
-			changeTexture = false;
-
-			(*sNumTextures)++;
-		}
-		if (changePolyFlags)
-		{
-			SetBlend(nextPolyFlags);
-			currentPolyFlags = nextPolyFlags;
-			changePolyFlags = false;
-
-			(*sNumPolyFlags)++;
-		}
-		if (changeSurfaceInfo)
-		{
-			GLRGBAFloat poly = {0,0,0,0};
-			GLRGBAFloat tint = {0,0,0,0};
-			GLRGBAFloat fade = {0,0,0,0};
-			gl_shaderprogramchanged = false;
-			if (nextPolyFlags & PF_Modulated)
-			{
-				// Poly color
-				poly.red    = byte2float[nextSurfaceInfo.PolyColor.s.red];
-				poly.green  = byte2float[nextSurfaceInfo.PolyColor.s.green];
-				poly.blue   = byte2float[nextSurfaceInfo.PolyColor.s.blue];
-				poly.alpha  = byte2float[nextSurfaceInfo.PolyColor.s.alpha];
-				pglColor4ubv((GLubyte*)&nextSurfaceInfo.PolyColor.s);
-			}
-			if (gl_allowshaders)
-			{
-				// Tint color
-				tint.red   = byte2float[nextSurfaceInfo.TintColor.s.red];
-				tint.green = byte2float[nextSurfaceInfo.TintColor.s.green];
-				tint.blue  = byte2float[nextSurfaceInfo.TintColor.s.blue];
-				tint.alpha = byte2float[nextSurfaceInfo.TintColor.s.alpha];
-				// Fade color
-				fade.red   = byte2float[nextSurfaceInfo.FadeColor.s.red];
-				fade.green = byte2float[nextSurfaceInfo.FadeColor.s.green];
-				fade.blue  = byte2float[nextSurfaceInfo.FadeColor.s.blue];
-				fade.alpha = byte2float[nextSurfaceInfo.FadeColor.s.alpha];
-
-				Shader_Load(&nextSurfaceInfo, &poly, &tint, &fade);
-				
-				if (gl_palshader)
-				{
-					pglActiveTexture(GL_TEXTURE2);// this stuff could be done better but gonna do it quick like this for now
-					pglBindTexture(GL_TEXTURE_2D, nextSurfaceInfo.LightTableId);
-					pglActiveTexture(GL_TEXTURE0);
-				}
-			}
-
-			currentSurfaceInfo = nextSurfaceInfo;
-			changeSurfaceInfo = false;
-
-			(*sNumColors)++;
-		}
-		// and that should be it?
-	}
-	// reset the arrays (set sizes to 0)
-	polygonArraySize = 0;
-	unsortedVertexArraySize = 0;
-
-	*sDrawTime = I_GetPreciseTime() - *sDrawTime;
+	Shader_Load(pSurf, &poly, &tint, &fade);
 }
 
 // -----------------+
 // DrawPolygon      : Render a polygon, set the texture, set render mode
 // -----------------+
-EXPORT void HWRAPI(DrawPolygon) (FSurfaceInfo *pSurf, FOutVector *pOutVerts, FUINT iNumPts, FBITFIELD PolyFlags, boolean horizonSpecial)
+EXPORT void HWRAPI(DrawPolygon) (FSurfaceInfo *pSurf, FOutVector *pOutVerts, FUINT iNumPts, FBITFIELD PolyFlags)
 {
-	if (gl_batching)
-	{
-		//CONS_Printf("Batched DrawPolygon\n");
-		if (!pSurf)
-			I_Error("Got a null FSurfaceInfo in batching");// nulls should only come in sky background pic drawing
-		if (polygonArraySize == polygonArrayAllocSize)
-		{
-			PolygonArrayEntry* new_array;
-			// ran out of space, make new array double the size
-			polygonArrayAllocSize *= 2;
-			new_array = malloc(polygonArrayAllocSize * sizeof(PolygonArrayEntry));
-			memcpy(new_array, polygonArray, polygonArraySize * sizeof(PolygonArrayEntry));
-			free(polygonArray);
-			polygonArray = new_array;
-			// also need to redo the index array, dont need to copy it though
-			free(polygonIndexArray);
-			polygonIndexArray = malloc(polygonArrayAllocSize * sizeof(unsigned int));
-		}
+	PreparePolygon(pSurf, pOutVerts, PolyFlags);
 
-		while (unsortedVertexArraySize + (int)iNumPts > unsortedVertexArrayAllocSize)
-		{
-			FOutVector* new_array;
-			// need more space for vertices in unsortedVertexArray
-			unsortedVertexArrayAllocSize *= 2;
-			new_array = malloc(unsortedVertexArrayAllocSize * sizeof(FOutVector));
-			memcpy(new_array, unsortedVertexArray, unsortedVertexArraySize * sizeof(FOutVector));
-			free(unsortedVertexArray);
-			unsortedVertexArray = new_array;
-		}
+	pglVertexPointer(3, GL_FLOAT, sizeof(FOutVector), &pOutVerts[0].x);
+	pglTexCoordPointer(2, GL_FLOAT, sizeof(FOutVector), &pOutVerts[0].s);
+	pglDrawArrays(GL_TRIANGLE_FAN, 0, iNumPts);
 
-		// add the polygon data to the arrays
+	if (PolyFlags & PF_RemoveYWrap)
+		pglTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_T, GL_REPEAT);
 
-		polygonArray[polygonArraySize].surf = *pSurf;
-		polygonArray[polygonArraySize].vertsIndex = unsortedVertexArraySize;
-		polygonArray[polygonArraySize].numVerts = iNumPts;
-		polygonArray[polygonArraySize].polyFlags = PolyFlags;
-		polygonArray[polygonArraySize].texNum = tex_downloaded;
-		polygonArray[polygonArraySize].shader = gl_currentshaderprogram;
-		polygonArray[polygonArraySize].horizonSpecial = horizonSpecial;
-		polygonArraySize++;
+	if (PolyFlags & PF_ForceWrapX)
+		Clamp2D(GL_TEXTURE_WRAP_S);
 
-		memcpy(&unsortedVertexArray[unsortedVertexArraySize], pOutVerts, iNumPts * sizeof(FOutVector));
-		unsortedVertexArraySize += iNumPts;
-	}
-	else
-	{
-		static GLRGBAFloat poly = {0,0,0,0};
-		static GLRGBAFloat tint = {0,0,0,0};
-		static GLRGBAFloat fade = {0,0,0,0};
+	if (PolyFlags & PF_ForceWrapY)
+		Clamp2D(GL_TEXTURE_WRAP_T);
+}
 
-		SetBlend(PolyFlags);    //TODO: inline (#pragma..)
+EXPORT void HWRAPI(DrawIndexedTriangles) (FSurfaceInfo *pSurf, FOutVector *pOutVerts, FUINT iNumPts, FBITFIELD PolyFlags, unsigned int *IndexArray)
+{
+	PreparePolygon(pSurf, pOutVerts, PolyFlags);
 
-		if (gl_portal_mode == HWD_PORTAL_NORMAL)
-		{
-			// PolyColor
-			if (pSurf)
-			{
-				// If modulated, mix the surface colour to the texture
-				if (CurrentPolyFlags & PF_Modulated)
-					pglColor4ubv((GLubyte*)&pSurf->PolyColor.s);
+	pglVertexPointer(3, GL_FLOAT, sizeof(FOutVector), &pOutVerts[0].x);
+	pglTexCoordPointer(2, GL_FLOAT, sizeof(FOutVector), &pOutVerts[0].s);
+	pglDrawElements(GL_TRIANGLES, iNumPts, GL_UNSIGNED_INT, IndexArray);
 
-				// If the surface is either modulated or colormapped, or both
-				if (CurrentPolyFlags & (PF_Modulated | PF_ColorMapped))
-				{
-					poly.red   = byte2float[pSurf->PolyColor.s.red];
-					poly.green = byte2float[pSurf->PolyColor.s.green];
-					poly.blue  = byte2float[pSurf->PolyColor.s.blue];
-					poly.alpha = byte2float[pSurf->PolyColor.s.alpha];
-				}
-
-				// Only if the surface is colormapped
-				if (CurrentPolyFlags & PF_ColorMapped)
-				{
-					tint.red   = byte2float[pSurf->TintColor.s.red];
-					tint.green = byte2float[pSurf->TintColor.s.green];
-					tint.blue  = byte2float[pSurf->TintColor.s.blue];
-					tint.alpha = byte2float[pSurf->TintColor.s.alpha];
-
-					fade.red   = byte2float[pSurf->FadeColor.s.red];
-					fade.green = byte2float[pSurf->FadeColor.s.green];
-					fade.blue  = byte2float[pSurf->FadeColor.s.blue];
-					fade.alpha = byte2float[pSurf->FadeColor.s.alpha];
-				}
-			
-				if (gl_palshader && gl_allowshaders)
-				{
-					pglActiveTexture(GL_TEXTURE2);
-					pglBindTexture(GL_TEXTURE_2D, pSurf->LightTableId);
-					pglActiveTexture(GL_TEXTURE0);
-				}
-			}
-
-		Shader_Load(pSurf, &poly, &tint, &fade);
-		}
-		else
-			UnSetShader();
-
-		pglVertexPointer(3, GL_FLOAT, sizeof(FOutVector), &pOutVerts[0].x);
-		pglTexCoordPointer(2, GL_FLOAT, sizeof(FOutVector), &pOutVerts[0].s);
-		pglDrawArrays(GL_TRIANGLE_FAN, 0, iNumPts);
-
-		if (PolyFlags & PF_RemoveYWrap)
-			pglTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_T, GL_REPEAT);
-
-		if (PolyFlags & PF_ForceWrapX)
-			Clamp2D(GL_TEXTURE_WRAP_S);
-
-		if (PolyFlags & PF_ForceWrapY)
-			Clamp2D(GL_TEXTURE_WRAP_T);
-	}
+	// the DrawPolygon variant of this has some code about polyflags and wrapping here but havent noticed any problems from omitting it?
 }
 
 // Sky dome code, taken/backported from SRB2
