+++ src/sdl/hwsym_sdl.c
@@ -81,6 +81,7 @@ void *hwSym(const char *funcName,void *handle)
 	GETFUNC(SetupGLInfo);
 	GETFUNC(Draw2DLine);
 	GETFUNC(DrawPolygon);
+	GETFUNC(DrawIndexedTriangles);
 	GETFUNC(SetBlend);
 	GETFUNC(ClearBuffer);
 	GETFUNC(SetTexture);
@@ -114,10 +115,7 @@ void *hwSym(const char *funcName,void *handle)
 	GETFUNC(SetShaderInfo);
 	GETFUNC(LoadCustomShader);
 	GETFUNC(InitCustomShaders);
-	
-	GETFUNC(StartBatching);
-	GETFUNC(RenderBatches);
-	
+
 	GETFUNC(InitPalette);
 	GETFUNC(AddLightTable);
 	GETFUNC(ClearLightTableCache);
