+++ src/sdl/i_video.c
@@ -2113,6 +2113,7 @@ void I_StartupGraphics(void)
 		HWD.pfnFinishUpdate     = NULL;
 		HWD.pfnDraw2DLine       = hwSym("Draw2DLine",NULL);
 		HWD.pfnDrawPolygon      = hwSym("DrawPolygon",NULL);
+		HWD.pfnDrawIndexedTriangles = hwSym("DrawIndexedTriangles",NULL);
 		HWD.pfnSetBlend         = hwSym("SetBlend",NULL);
 		HWD.pfnClearBuffer      = hwSym("ClearBuffer",NULL);
 		HWD.pfnSetTexture       = hwSym("SetTexture",NULL);
@@ -2147,9 +2148,6 @@ void I_StartupGraphics(void)
 		HWD.pfnSetShaderInfo    = hwSym("SetShaderInfo",NULL);
 		HWD.pfnLoadCustomShader = hwSym("LoadCustomShader",NULL);
 		HWD.pfnInitCustomShaders = hwSym("InitCustomShaders",NULL);
-
-		HWD.pfnStartBatching = hwSym("StartBatching",NULL);
-		HWD.pfnRenderBatches = hwSym("RenderBatches",NULL);
 
 		HWD.pfnInitPalette = hwSym("InitPalette",NULL);
 		HWD.pfnAddLightTable = hwSym("AddLightTable",NULL);
